requirements.txt
Django==4.2.1
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
python-telegram-bot==20.3
django-tinymce==3.6.1
djangorestframework-simplejwt==5.2.2
